requirements.txt
streamlit
pandas
folium
streamlit-folium
geopy
requests
Pillow